Extract the calculations methods into a module
