+++ models/committee.rb
@@ -1,18 +1,5 @@
 class Committee < ActiveRecord::Base
-
-  has_many :calculations, as: :subject
-
-  def calculation(name)
-    @_calculations_cache ||= calculations.index_by(&:name)
-    @_calculations_cache[name.to_s].try(:value)
-  end
-
-  def save_calculation(name, value)
-    calculations
-      .where(name: name)
-      .first_or_create
-      .update_attributes(value: value)
-  end
+  include HasCalculations
 
   def as_json(options = nil)
     {
